# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RiceBook API",
    "version": "1.0.0",
    "description": "RiceBook 后端 API 文档"
  },
  "paths": {
...
    "/following/{user}": {
      "get": {
        "summary": "Fetch the list of users a user is following",
        "description": "Retrieve the list of usernames the specified user is following. If no username is provided, retrieve the list for the logged-in user.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "The username of the user"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the following list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "following": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "summary": "Add a user to the following list",
        "description": "Add the specified user to the logged-in user's following list.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The username of the user to follow"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully added to the following list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "following": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Username to follow is required"
          },
          "404": {
            "description": "User to follow not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "summary": "Remove a user from the following list",
        "description": "Remove the specified user from the logged-in user's following list.",
        "parameters": [
          {
            "name": "user",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The username of the user to unfollow"
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully removed from the following list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "following": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Username to unfollow is required"
          },
          "404": {
            "description": "User to unfollow not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}